sync: Convert TabNodePool to use tab_node_id as primary index.
The tab node id and sync id are similarly unique for purposes of
TabNodePool, and having both is functionally redundant.  Since the
Sync API is designed to work with client tag based referencing and
not IDs, and we want to convert sessions to use the sync API, this
patch removes use of sync IDs as the primary lookup key in
TabNodePool. 

chrome/android & chrome/browser/android OWNERS:
TBR=yfriedman@chromium.org


BUG=98892,80194

Review URL: https://chromiumcodereview.appspot.com/19529003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@215208 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: a4d45a029ae0eb86e1a142f6b5b12c5d2f628aaf

+++ chrome/android/testshell/testshell_tab.cc
@@ -131,9 +131,9 @@ static jint Init(JNIEnv* env,
   return reinterpret_cast<jint>(tab);
 }
 
-int64 TestShellTab::GetSyncId() const {
+int TestShellTab::GetSyncId() const {
   NOTIMPLEMENTED();
   return 0;
 }
 
-void TestShellTab::SetSyncId(int64 sync_id) { NOTIMPLEMENTED(); }
+void TestShellTab::SetSyncId(int sync_id) { NOTIMPLEMENTED(); }
