+++ chrome/browser/sync/glue/synced_tab_delegate.h
@@ -55,8 +55,8 @@ class SyncedTabDelegate {
   virtual bool HasWebContents() const = 0;
 
   // Session sync related methods.
-  virtual int64 GetSyncId() const = 0;
-  virtual void SetSyncId(int64 sync_id) = 0;
+  virtual int GetSyncId() const = 0;
+  virtual void SetSyncId(int sync_id) = 0;
   // Returns the SyncedTabDelegate associated with WebContents.
   static SyncedTabDelegate* ImplFromWebContents(
       content::WebContents* web_contents);
