+++ chrome/browser/sync/glue/synced_tab_delegate_android.h
@@ -38,8 +38,8 @@ class SyncedTabDelegateAndroid : public browser_sync::SyncedTabDelegate {
   virtual content::NavigationEntry* GetActiveEntry() const OVERRIDE;
   virtual bool IsPinned() const OVERRIDE;
   virtual bool HasWebContents() const OVERRIDE;
-  virtual int64 GetSyncId() const OVERRIDE;
-  virtual void SetSyncId(int64 sync_id) OVERRIDE;
+  virtual int GetSyncId() const OVERRIDE;
+  virtual void SetSyncId(int sync_id) OVERRIDE;
 
   // Managed user related methods.
