+++ chrome/browser/ui/sync/tab_contents_synced_tab_delegate.cc
@@ -101,10 +101,10 @@ bool TabContentsSyncedTabDelegate::IsPinned() const {
 
 bool TabContentsSyncedTabDelegate::HasWebContents() const { return true; }
 
-int64 TabContentsSyncedTabDelegate::GetSyncId() const {
+int TabContentsSyncedTabDelegate::GetSyncId() const {
   return sync_session_id_;
 }
 
-void TabContentsSyncedTabDelegate::SetSyncId(int64 sync_id) {
+void TabContentsSyncedTabDelegate::SetSyncId(int sync_id) {
   sync_session_id_ = sync_id;
 }
