+++ chrome/browser/ui/sync/tab_contents_synced_tab_delegate.h
@@ -37,15 +37,15 @@ class TabContentsSyncedTabDelegate
       GetBlockedNavigations() const OVERRIDE;
   virtual bool IsPinned() const OVERRIDE;
   virtual bool HasWebContents() const OVERRIDE;
-  virtual int64 GetSyncId() const OVERRIDE;
-  virtual void SetSyncId(int64 sync_id) OVERRIDE;
+  virtual int GetSyncId() const OVERRIDE;
+  virtual void SetSyncId(int sync_id) OVERRIDE;
 
  private:
   explicit TabContentsSyncedTabDelegate(content::WebContents* web_contents);
   friend class content::WebContentsUserData<TabContentsSyncedTabDelegate>;
 
   content::WebContents* web_contents_;
-  int64 sync_session_id_;
+  int sync_session_id_;
 
   DISALLOW_COPY_AND_ASSIGN(TabContentsSyncedTabDelegate);
 };
